[PMove]: Added back in support for "currents". [PMove/Utils]: More work on "events". :-)

+++ src/baseq2rtxp/sharedgame/sg_misc.cpp
@@ -36,9 +36,9 @@ const char *sg_player_state_event_strings[ PS_EV_MAX ] = {
 /**
 *	@brief	Adds a player movement predictable event to the player move state.
 **/
-void SG_PlayerState_AddPredictableEvent( const uint8_t newEvent, const uint8_t eventParm, player_state_t *playerState ) {
+void SG_PlayerState_AddPredictableEvent( const uint32_t newEvent, const uint32_t eventParm, player_state_t *playerState ) {
 	// Event sequence index.
-	const int32_t sequenceIndex = playerState->eventSequence & ( MAX_PS_EVENTS - 1 );
+	const int64_t sequenceIndex = playerState->eventSequence & ( MAX_PS_EVENTS - 1 );
 
 	// Ensure it is within bounds.
 	if ( newEvent < PS_EV_NONE || newEvent >= PS_EV_MAX ) {
