+++ src/baseq2rtxp/sharedgame/sg_misc.h
@@ -18,7 +18,7 @@
 /**
 *	@brief	Adds a player movement predictable event to the player move state.
 **/
-void SG_PlayerState_AddPredictableEvent( const uint8_t newEvent, const uint8_t eventParm, player_state_t *playerState );
+void SG_PlayerState_AddPredictableEvent( const uint32_t newEvent, const uint32_t eventParm, player_state_t *playerState );
